requirements.txt
setuptools>=67.0.0
wheel
Flask==3.0.3
google-api-python-client==1.12.8
python-dotenv

